Split up original import mapping and address PR comments.
