+++ src/actions/pause/mapScopes/index.js
@@ -4,36 +4,41 @@
 
 // @flow
 
-import { getSource } from "../../selectors";
-import { loadSourceText } from "../sources/loadSourceText";
+import { getSource } from "../../../selectors";
+import { loadSourceText } from "../../sources/loadSourceText";
 import {
   getScopes,
   type SourceScope,
   type BindingData,
-  type BindingLocation,
-  type BindingMetaValue,
-  type BindingType
-} from "../../workers/parser";
-import type { RenderableScope } from "../../utils/pause/scopes/getScope";
-import { PROMISE } from "../utils/middleware/promise";
+  type BindingLocation
+} from "../../../workers/parser";
+import type { RenderableScope } from "../../../utils/pause/scopes/getScope";
+import { PROMISE } from "../../utils/middleware/promise";
+import { locColumn } from "./utils";
 
-import { features } from "../../utils/prefs";
-import { log } from "../../utils/log";
+// eslint-disable-next-line max-len
+import { findGeneratedBindingFromPosition } from "./findGeneratedBindingFromPosition.js";
+
+import { features } from "../../../utils/prefs";
+import { log } from "../../../utils/log";
 import { isGeneratedId } from "devtools-source-map";
 import type {
   Frame,
   Scope,
   Source,
-  Location,
   BindingContents,
   ScopeBindings
-} from "../../types";
+} from "../../../types";
 
-import { createObjectClient } from "../../client/firefox";
-
-import type { ThunkArgs } from "../types";
+import type { ThunkArgs } from "../../types";
 
 export type OriginalScope = RenderableScope;
+
+export type GeneratedBindingLocation = {
+  name: string,
+  loc: BindingLocation,
+  desc: BindingContents | null
+};
 
 export function mapScopes(scopes: Promise<Scope>, frame: Frame) {
   return async function({ dispatch, getState, client, sourceMaps }: ThunkArgs) {
@@ -284,174 +289,6 @@ async function findGeneratedBinding(
   };
 }
 
-async function findGeneratedBindingFromPosition(
-  sourceMaps: any,
-  client: any,
-  source: Source,
-  pos: BindingLocation,
-  name: string,
-  type: BindingType,
-  generatedAstBindings: Array<GeneratedBindingLocation>
-) {
-  const gen = await sourceMaps.getGeneratedLocation(pos.start, source);
-  const genEnd = await sourceMaps.getGeneratedLocation(pos.end, source);
-
-  // Since the map takes the closest location, sometimes mapping a
-  // binding's location can point at the start of a binding listed after
-  // it, so we need to make sure it maps to a location that actually has
-  // a size in order to avoid picking up the wrong descriptor.
-  if (gen.line === genEnd.line && gen.column === genEnd.column) {
-    return null;
-  }
-
-  return generatedAstBindings.reduce(async (acc, val) => {
-    const accVal = await acc;
-    if (accVal) {
-      return accVal;
-    }
-
-    const operations = mapBindingToGeneratedRange(
-      val,
-      {
-        start: gen,
-        end: genEnd
-      },
-      type === "import"
-    );
-
-    if (operations) {
-      const hasCall = operations.some(op => op.type === "call");
-      if (hasCall) {
-        // In theory this could fall back to trying to use client.evaluate.
-        return null;
-      }
-
-      let desc = val.desc;
-      if (desc) {
-        for (const op of operations) {
-          if (op.type === "call" || op.type === "inherit") {
-            continue;
-          }
-
-          const objectClient = createObjectClient(desc.value);
-          desc = (await objectClient.getProperty(op.property)).descriptor;
-          if (!desc) {
-            return null;
-          }
-        }
-      }
-
-      return {
-        name: val.name,
-        desc: desc
-      };
-    }
-
-    return null;
-  }, null);
-}
-
-function mapBindingToGeneratedRange(
-  binding: GeneratedBindingLocation,
-  mapped: {
-    start: Location,
-    end: Location
-  },
-  allowExpressionMatch: boolean
-): Array<BindingMetaValue> | null {
-  // Allow the mapping to point anywhere within the generated binding
-  // location to allow for less than perfect sourcemaps. Since you also
-  // need at least one character between identifiers, we also give one
-  // characters of space at the front the generated binding in order
-  // to increase the probability of finding the right mapping.
-  if (
-    !allowExpressionMatch &&
-    mapped.start.line === binding.loc.start.line &&
-    locColumn(mapped.start) >= locColumn(binding.loc.start) - 1 &&
-    locColumn(mapped.start) <= locColumn(binding.loc.end)
-  ) {
-    return [];
-  }
-
-  if (allowExpressionMatch) {
-    // Expression matches require broader searching because sourcemaps usage
-    // varies in how they map certain things. For instance given
-    //
-    //   import { bar } from "mod";
-    //   bar();
-    //
-    // The "bar()" expression is generally expanded into one of two possibly
-    // forms, both of which map the "bar" identifier in different ways. See
-    // the "^^" markers below for the ranges.
-    //
-    //   (0, foo.bar)()    // Babel
-    //       ^^^^^^^       // mapping
-    //       ^^^           // binding
-    // vs
-    //
-    //   Object(foo.bar)() // Webpack
-    //   ^^^^^^^^^^^^^^^   // mapping
-    //          ^^^        // binding
-    //
-    // Unfortunately, Webpack also has a tendancy to over-map past the call
-    // expression to the start of the next line, at least when there isn't
-    // anything else on that line that is mapped, e.g.
-    //
-    //   Object(foo.bar)()
-    //   ^^^^^^^^^^^^^^^^^
-    //   ^                 // wrapped to column 0 of next line
-
-    if (mappingContains(mapped, binding.loc)) {
-      const { meta } = binding.loc;
-
-      // Limit to 2 simple property or inherits operartions, since it would
-      // just be more work to search mroe and it is very unlikely that
-      // bindings would be mapped to more than a single member + inherits
-      // wrapper.
-      const operations = [];
-      for (let op = meta, i = 0; op && i < 2; i++, op = op.parent) {
-        // Allow calls as the first access, "foo().bar", not but "foo.bar()"
-        // because that is too likely to accidentally pick up a call that
-        // was in the user's own code, and it would be too risky. The chance
-        // of an expression mapping "one" to "_one()" isn't a common transform
-        // and is also one that is less likely to be screwed up during
-        // sourcemap generation, so it is allowed.
-        if (op.type === "call" && i !== 0) {
-          break;
-        }
-
-        if (mappingContains(mapped, op)) {
-          if (op.type !== "inherit") {
-            operations.push(op);
-          }
-        } else {
-          break;
-        }
-      }
-
-      return operations;
-    }
-  }
-  return null;
-}
-
-function mappingContains(mapped, item) {
-  return (
-    (item.start.line > mapped.start.line ||
-      (item.start.line === mapped.start.line &&
-        locColumn(item.start) >= locColumn(mapped.start))) &&
-    (item.end.line < mapped.end.line ||
-      (item.end.line === mapped.end.line &&
-        locColumn(item.end) <= locColumn(mapped.end)))
-  );
-}
-
-type GeneratedBindingLocation = {
-  name: string,
-  loc: BindingLocation,
-  desc: BindingContents | null
-};
-
 function buildGeneratedBindingList(
   scopes: Scope,
   generatedAstScopes: SourceScope[],
@@ -522,14 +359,4 @@ function buildGeneratedBindingList(
     });
 
   return generatedBindings;
-}
-
-function locColumn(loc: Location): number {
-  if (typeof loc.column !== "number") {
-    // This shouldn't really happen with locations from the AST, but
-    // the datatype we are using allows null/undefined column.
-    return 0;
-  }
-
-  return loc.column;
 }
